backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Environment management
python-dotenv==1.0.0

# OpenAI API
openai==1.3.7

# Document processing
docling>=1.0.0

# File handling and utilities
Werkzeug==3.0.1

# Logging and monitoring
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0 